created new plugin partial finder
